fix: consider chainId change in useIsBlockNumberRelevant
