+++ libs/balances-and-allowances/src/hooks/useNativeTokensBalances.ts
@@ -1,15 +1,18 @@
 import { useMemo } from 'react'
 
-import { usePrevious } from '@cowprotocol/common-hooks'
+import { SupportedChainId } from '@cowprotocol/cow-sdk'
 import { getMulticallContract, useSingleContractMultipleData } from '@cowprotocol/multicall'
 import { useWalletProvider } from '@cowprotocol/wallet-provider'
 import { BigNumber } from '@ethersproject/bignumber'
 
-import { getIsBlockNumberRelevant } from '../utils/getIsBlockNumberRelevant'
+import { useIsBlockNumberRelevant } from './useIsBlockNumberRelevant'
 
 type NativeBalances = { [account: string]: BigNumber | undefined }
 
-export function useNativeTokensBalances(accounts: string[] | undefined): NativeBalances | undefined {
+export function useNativeTokensBalances(
+  chainId: SupportedChainId,
+  accounts: string[] | undefined,
+): NativeBalances | undefined {
   const provider = useWalletProvider()
   const contract = provider ? getMulticallContract(provider) : undefined
   const params = useMemo(() => accounts?.map((account) => [account]), [accounts])
@@ -17,10 +20,9 @@ export function useNativeTokensBalances(accounts: string[] | undefined): NativeB
   const { data } = useSingleContractMultipleData<[BigNumber]>(contract, 'getEthBalance', params)
   const results = data?.results
   const blockNumber = data?.blockNumber
-  const prevBlockNumber = usePrevious(blockNumber)
 
   // Skip multicall results from outdated block if there is a result from newer one
-  const isNewBlockNumber = getIsBlockNumberRelevant({ prevBlockNumber, blockNumber })
+  const isNewBlockNumber = useIsBlockNumberRelevant(chainId, blockNumber)
 
   return useMemo(() => {
     if (!results || !accounts || !isNewBlockNumber) return undefined
