--- libs/balances-and-allowances/src/utils/getIsBlockNumberRelevant.ts
@@ -1,9 +1,0 @@
-export function getIsBlockNumberRelevant<T = number | undefined>({
-  blockNumber,
-  prevBlockNumber,
-}: {
-  blockNumber: T
-  prevBlockNumber: T
-}): boolean {
-  return typeof prevBlockNumber === 'number' && typeof blockNumber === 'number' ? blockNumber > prevBlockNumber : true
-}
